@@ -37,7 +37,7 @@ module ParamsDeserializers
       deserialized_params_name = options.delete(:as).try(:to_sym) || :deserialized_params
       attr_reader deserialized_params_name
 
-      before_filter(options) do
+      before_action(options) do
         instance_variable_set("@#{deserialized_params_name}", deserializer.new(params).deserialize)
       end
     end
