@@ -37,11 +37,10 @@ describe 'deserialize_params_with', type: :controller do
       deserializer = Class.new(ParamsDeserializer) do
         attributes :foo
       end
-
       deserialize_params_with deserializer, only: :update
 
       def update
-        render text: ''
+        render plain: ''
       end
     end
 
@@ -54,7 +53,7 @@ describe 'deserialize_params_with', type: :controller do
 
     it 'gets deserialized params when calling the deserialized_params getter' do
       routes.draw { get 'update' => 'anonymous#update' }
-      put :update, foo: 'bar', baz: 'quux'
+      put :update, params: { foo: 'bar', baz: 'quux' }
 
       expected = { foo: 'bar' }.with_indifferent_access
       expect(controller.deserialized_params).to eql expected
@@ -72,7 +71,7 @@ describe 'deserialize_params_with', type: :controller do
       deserialize_params_with deserializer, as: :deserialized_params_foo, only: :update
 
       def update
-        render text: ''
+        render plain: ''
       end
     end
 
