@@ -89,7 +89,7 @@ class ParamsDeserializer
   end
 
   def verify_valid_keys
-    invalid_params = params_root.symbolize_keys.keys - self.class.attrs.map(&:original_name)
+    invalid_params = params.permit(params_root.keys).to_h.symbolize_keys.keys - self.class.attrs.map(&:original_name)
     if self.class.strict_mode && !invalid_params.empty?
       raise InvalidKeyError, "Invalid keys in params: #{invalid_params.map(&:inspect).join(",")}."
     end
